+++ src/main/java/blackjack/domain/card/CardType.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum CardType {
     HEART("하트"),
