+++ src/test/java/blackjack/PlayerTest.java
@@ -1,9 +1,9 @@
 package blackjack;
 
-import blackjack.domain.Card;
-import blackjack.domain.CardNumber;
-import blackjack.domain.CardType;
-import blackjack.domain.Player;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardType;
+import blackjack.domain.participants.Player;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
